refactor: change type data double to int for all atribut

+++ src/Character.java
@@ -1,6 +1,6 @@
 public class Character {
     protected String name;
-    protected double health;
-    protected double damage;
-    protected double armor;
+    protected int health;
+    protected int damage;
+    protected int armor;
 }
